==> 03_basic_preprocessed_data_output.ipynb <==
## File 03 - Basic Preprocessed Data Output

In this file, we create a small preprocessed data table.

### Set up Spark session and data schema

We can specify more options in the SparkSession creator, but currently the options are at the default settings.

%%time
from pyspark.sql import SparkSession
from pyspark.sql import types as T

spark = SparkSession.builder \
        .appName("project") \
        .getOrCreate()

sc = spark.sparkContext

#schema = "`event_time` TIMESTAMP,`event_type` STRING,`product_id` INT,`category_id` BIGINT,`category_code` STRING,`brand` STRING,`price` FLOAT,`user_id` INT,`user_session` STRING"
#ddl_schema = T._parse_datatype_string(schema)

### Read in dataframes

%%time
df1 = spark.read.parquet("/project/ds5559/group12/raw_data/FilteredJan2020M1.parquet")
df2 = spark.read.parquet("/project/ds5559/group12/raw_data/FilteredFeb2020M2.parquet")

### Limit number of records in dataframes

We can limit each dataframe to a smaller subset. Notably, the dataframe is arranged by time, so this is how the subset will be biased.

# df1=df1.limit(100000)
df1.createOrReplaceTempView("m1")

# df2=df11.limit(100000)
df2.createOrReplaceTempView("m2")

### Create a transformed table containing elements of interest for our model

We propose a basic table format (see https://docs.google.com/document/d/1NG4KGticBXn0D3PL5_zMxLV2Pr7A8PQtLcasxCOd1nA/edit).

Every row is a user_id who exists in M1 and may or may not exist in M2.
Columns include:
- `user_id` (ID)
- `m2_total_spend` (sum among all purchase events, NB: month 2. Will be 0 is user does not exist in month 2 or makes no purchase events)
- `m1_total_spend` (sum among all purchase events)
- `m1_user_sessions` (count of distinct user sessions/browsing sessions)
- `m1_purchase_events` (count of distinct purchase events)

Additional rows may have been added below.


%%time
df = spark.sql("""SELECT

               /* ID */
               m1.user_id AS user_id,
               
               /* Output: Total spend in month 2; Will be 0 is user does not exist in month 2 or makes no purchase events */
               IFNULL(SUM(m2.price),0) AS m2_total_spend,
               
               
               /* Total spend in month 1: we sum the price of any 'purchase' events */
               SUM(CAST(m1.event_type=='purchase' AS INT) * m1.price) AS m1_total_spend,
               
               /* Total events in month 1: includes all event types (multiple per session) */
               COUNT(m1.event_type) AS m1_total_events,
               
               /* Total purchase events in month 1: we sum the occurence of any 'purchase' events */
               SUM(CAST(m1.event_type=='purchase' AS INT)) AS m1_purchase_events,
               
               /* Total user sessions in month 1: we count all distinct user sessions */
               COUNT(DISTINCT m1.user_session) AS m1_user_sessions
               
               FROM m1
               
            LEFT JOIN 
            (
               SELECT * FROM m2
               WHERE event_type='purchase'
            ) m2
            
            ON m1.user_id=m2.user_id
            
            GROUP BY m1.user_id ORDER BY m1_purchase_events DESC""")
df.show(50)

#### Write to file. Note that if you create a file here, you must run chmod 777 filename in the folder in bash to give all access

%%time
df.write.mode("overwrite").parquet("/project/ds5559/group12/processed_data/output_from_03.parquet")
