==> 04_eda2.ipynb <==
## File 04 - Exploratory Data Analysis, Month to Month Level
##### Group 12:

##### Hannah Schmuckler, mmc4cv

##### Rob Schwartz, res7cd

In this file, we look into comparing users and their characteristics over two months. The months we chose are January and February 2020. For the purposes of our analysis, month 1 is January and Month 2 is February. We are predicting spend in month 2 based on characteristics from month 1. 

### Set up Spark session and data schema

We can specify more options in the SparkSession creator, but currently the options are at the default settings.

%%time
from pyspark.sql import SparkSession
from pyspark.sql import types as T

spark = SparkSession.builder \
        .appName("project") \
        .getOrCreate()

sc = spark.sparkContext

schema = "`event_time` TIMESTAMP,`event_type` STRING,`product_id` INT,`category_id` BIGINT,`category_code` STRING,`brand` STRING,`price` FLOAT,`user_id` INT,`user_session` STRING"
ddl_schema = T._parse_datatype_string(schema)

### Read in dataframes for two months

%%time
df1 = spark.read.schema(ddl_schema).csv("/project/ds5559/group12/raw_data/2020-01.csv")
df2 = spark.read.schema(ddl_schema).csv("/project/ds5559/group12/raw_data/2020-02.csv")

### Limit number of records in dataframes

We can limit each dataframe to a smaller subset. Notably, the dataframe is arranged by time, so this is how the subset will be biased.

# df1=df1.limit(10000)
df1.createOrReplaceTempView("r1")

# df2=df2.limit(10000)
df2.createOrReplaceTempView("r2")

### See how many users are the same

##### Full dataset

- Over all interactions from each month: 1,702,723 are the same users
- This is out of 4,385,986 users in month 1 and 4,233,207 users in month 2
- So about 38% to 40% of users are the same from month to month

%%time
spark.sql("SELECT DISTINCT r1.user_id FROM r1 INNER JOIN r2 on r1.user_id=r2.user_id").count()

%%time
spark.sql("SELECT DISTINCT r1.user_id FROM r1").count()

%%time
spark.sql("SELECT DISTINCT r2.user_id FROM r2").count()

### See how many users made purchases in both months

##### Full dataset

- Over all interactions from each month: 93,209 are the same purchasers
- This is out of 359,105 purchasers in month 1 and 392,356 purchasers in month 2
- This is out of 4,385,986 users in month 1 and 4,233,207 users in month 2

- So about 24%-26% of purchasing users are the same from month to month
- And about 2% of any users purchase in both months


%%time
spark.sql("""SELECT DISTINCT r1.user_id FROM r1 INNER JOIN r2 on r1.user_id=r2.user_id WHERE r1.event_type="purchase" and r2.event_type="purchase" """).count()

%%time
spark.sql("""SELECT DISTINCT r1.user_id FROM r1 WHERE r1.event_type="purchase" """).count()

%%time
spark.sql("""SELECT DISTINCT r2.user_id FROM r2 WHERE r2.event_type="purchase" """).count()

### See some similar user behavior

Let's look at the similarity of products purchased between users in each month. Takes about 1m30s to run.

We can see that many products purchased in Month 10 are in the same category as products purchased in Month 11. Lots of nulls tend to clog up the dataset, however.

%%time
spark.sql("""SELECT uid, "1" AS month, category_code, event_type FROM (
             SELECT DISTINCT r1.user_id AS uid FROM r1 INNER JOIN r2 ON r1.user_id=r2.user_id WHERE r1.event_type="purchase" and r2.event_type="purchase"
              ) LEFT JOIN r1 ON uid=r1.user_id WHERE r1.event_type="purchase"
              
              UNION ALL
              
              SELECT uid, "2" AS month, category_code, event_type FROM (
              SELECT DISTINCT r1.user_id AS uid FROM r1 INNER JOIN r2 ON r1.user_id=r2.user_id WHERE r1.event_type="purchase" and r2.event_type="purchase"
              ) LEFT JOIN r2 ON uid=r2.user_id WHERE r2.event_type="purchase"
              
              ORDER BY uid, month ASC
              
              """).show(10,False)
# spark.sql("SELECT DISTINCT r1.user_id FROM r1 INNER JOIN r2 on r1.user_id=r2.user_id").count()

# %%time

# spark.sql("DROP TABLE IF EXISTS r_all")
# spark.sql("CREATE TABLE r_all LIKE r1").count()
# spark.sql("INSERT INTO r_all TABLE r1")
# spark.sql("INSERT INTO r_all TABLE r2")
# spark.sql("SELECT * FROM r_all").count()
