==> 08_two_part.ipynb <==
## File 08 - Two-part model (Logistic + Linear Regression)

##### Group 12:

##### Hannah Schmuckler, mmc4cv

##### Rob Schwartz, res7cd

In this file, we create a small ML pipeline based on the output from File 02 (Feature creation).

We create a two part model made up of a logistic and a linear regression.

### Set up Spark session

We can specify more options in the SparkSession creator, but currently the options are at the default settings.

%%time
from pyspark.sql import SparkSession
from pyspark.sql import types as T
from pyspark.ml import Pipeline
from pyspark.ml.feature import VectorAssembler, StandardScaler
from pyspark.ml.regression import LinearRegression, LinearRegressionModel, LinearRegressionSummary
from pyspark.ml.evaluation import RegressionEvaluator, BinaryClassificationEvaluator
from pyspark.ml.tuning import ParamGridBuilder, CrossValidator
from pyspark.sql.functions import col
from pyspark.sql.functions import log
from pyspark.ml.stat import Correlation
from pyspark.sql.functions import when
from pyspark.ml.classification import LogisticRegression
from handyspark import *
from matplotlib import pyplot as plt
from pyspark.mllib.evaluation import BinaryClassificationMetrics, MulticlassMetrics
from pyspark.sql.types import IntegerType, DoubleType
from pyspark.sql.functions import lit

import pandas as pd
import numpy as np
import copy

spark = SparkSession.builder \
        .appName("project") \
        .getOrCreate()

sc = spark.sparkContext

### Read in dataframes for train and test sets

This data should have been previously generated: we can find it in the `processed_data` folder.

%%time
trainDF = spark.read.parquet("./processed_data/train.parquet")
testDF = spark.read.parquet("./processed_data/test.parquet")
# trainDF.show(5)

### Label trainDF with 0 if they did not spend in m2 and 1 if they did

trainDF = trainDF.withColumn('logistic_T_total_spend', when(trainDF.T_total_spend == 0, 0)
                                                       .otherwise(1))
testDF = testDF.withColumn('logistic_T_total_spend', when(testDF.T_total_spend == 0, 0)
                                                       .otherwise(1))

trainDF = trainDF.withColumn('logistic_T_total_spend', col("logistic_T_total_spend").cast(DoubleType()))
testDF = testDF.withColumn('logistic_T_total_spend', col("logistic_T_total_spend").cast(DoubleType()))

### Set up Spark ML pipeline training for logistic regression

Here we decide which input columns should be used in order to create our training pipeline. To implement this step, we create the function `generatePipeline(inputCols, outputCol)`. Then, we train the pipeline using this function.

%%time

def generateLogisticPipeline(inputCols, outputCol, threshold):
    # Select input columns for linear regression
    vecAssembler = VectorAssembler(inputCols=inputCols, outputCol="unscaled_features")
    # Standard Scale
    ss = StandardScaler(inputCol="unscaled_features", outputCol="features")
    # Select output column for linear regression
    logR = LogisticRegression(featuresCol="features", labelCol=outputCol, rawPredictionCol = "Logistic_Probabilities", predictionCol = "Logistic_Predictions", threshold = threshold)

    # The following lines (pipeline creation and fitting) replace these two commented-out lines.
    # vecTrainDF = vecAssembler.transform(trainDF)
 
    pipeline = Pipeline(stages=[vecAssembler, ss, logR])
    return pipeline
    

### View the model information

Create function to print out the model coefficients for the overall model, RMSE and R^2. We define the functions `modelInfo(inputCols, pipelineModel)` and `getEvaluationMetrics(pipelineModel,outputCol,testDF)` to report this information.

def modelInfo(inputCols, pipelineModel):
    # Create a zipped list containing the coefficients and the data
    modelCols = copy.deepcopy(inputCols)
    modelCoeffs = list(pipelineModel.stages[-1].coefficients)
    modelCoeffs.insert(0,pipelineModel.stages[-1].intercept)
    modelCols.insert(0,"intercept")
    modelZippedList = list(map(list, zip(modelCols, modelCoeffs)))
    
    
    # Create the pandas DataFrame
    modelDF = pd.DataFrame(modelZippedList, columns = ['Column name', 'Coefficient'])
    #modelDF['pValues'] = pvals
    return modelDF

# Calculate adjusted r2 (https://towardsdatascience.com/machine-learning-linear-regression-using-pyspark-9d5d5c772b42)
# This function will allow us to calculate the adjusted r-square value when we do PCA later. The default r-square function does not take into account k. 
def adj_r2(r2, inputCols, testDF, k = 0):
    n = testDF.count()
    if k == 0:
        p = len(inputCols)
    else: 
        p = len(inputCols) + k - 1
    
    adjusted_r2 = 1-(((1-r2)*(n-1))/(n-p-1))
    return adjusted_r2

# https://towardsdatascience.com/binary-classifier-evaluation-made-easy-with-handyspark-3b1e69c12b4f
def getLogisticEvaluationMetrics(pipelineModel,outputCol,testDF,inputCols):
    predDF = pipelineModel.transform(testDF)
    preds_only = predDF.select(outputCol, "Logistic_Predictions")
    preds_only.show(5)
    
    preds_only = preds_only.rdd
    
    metrics = MulticlassMetrics(preds_only)
    print(metrics.confusionMatrix().toArray())
    print("Accuracy: " + str(metrics.accuracy))

    return predDF


## Part 1: Logistic regression

%%time
inputCols = ["total_spend", "total_events", "total_sessions", "avg_session_length", "sd_session_length", "avg_interactions_per_session", 
             "sd_interactions_per_session", "max_interactions_per_session", "purchase_pct_of_total_events", "view_pct_of_total_events",
             "cart_pct_of_total_events", "avg_purchases_per_session", "sd_purchases_per_session", "cart_events", "purchase_events", 
             "sessions_with_purchase", "sessions_with_cart", "sessions_with_view", "pct_sessions_end_purchase", 
             "pct_sessions_end_cart", "total_spend_log", "total_events_log", "purchase_events_log", "total_sessions_log",
             "avg_session_length_log"]
outputCol= "logistic_T_total_spend"

pipeline = generateLogisticPipeline(inputCols, outputCol, .2) #Iterative testing determined that a threshold of .2 led to the highest overal adj R^2 for the model
pipelineModel = pipeline.fit(trainDF)

print(modelInfo(inputCols, pipelineModel))

logistic_predictions = getLogisticEvaluationMetrics(pipelineModel,outputCol,testDF, inputCols)

#### Threshold of .2 maximizes model adj-r^2, though it does not maximize the accuracy of this logistic regression. 

## Part 2: Linear Regression

#### Next, we'll train a linear regression model using only the true positive training data, and generate a regression prediction for all test data that had a prediction of positive. 

part_2_train = trainDF.filter(col("logistic_T_total_spend") == 1)
part_2_test = logistic_predictions.filter(col("Logistic_Predictions") == 1).withColumnRenamed('probability', "ProbabilityLogistic")

part_1_test = logistic_predictions.filter(col("Logistic_Predictions") == 0).withColumnRenamed('probability', "ProbabilityLogistic") \
    .withColumn("", lit(None)) ## For overall accuracy metrics later, will be joined with output of part_2_test's regression predictions

print(part_2_train.count())
print(part_2_test.count())

%%time

def generateLinearPipeline(inputCols, outputCol):
    # Select input columns for linear regression
    vecAssembler = VectorAssembler(inputCols=inputCols, outputCol="unscaled_features2")
    # Standard Scale
    ss = StandardScaler(inputCol="unscaled_features2", outputCol="features2")
    # Select output column for linear regression
    lr = LinearRegression(featuresCol="features2", labelCol=outputCol)

    # The following lines (pipeline creation and fitting) replace these two commented-out lines.
    # vecTrainDF = vecAssembler.transform(trainDF)
 
    pipeline = Pipeline(stages=[vecAssembler, ss, lr])
    return pipeline
    
pipeline = generateLinearPipeline(inputCols, "T_total_spend")
pipelineModel = pipeline.fit(trainDF)



def getLinearEvaluationMetrics(linearPipelineModel,outputCol,testDF,inputCols):
    new_preds = pipelineModel.transform(testDF)
    
    # Merge the predictions from the linear regression with predictions from the logistic regression performed above. 
    # That will allow this to output complete results. 
    predDF = new_preds.unionByName(part_1_test, allowMissingColumns=True)

    predDF = predDF.withColumn("FinalPredictions", when(col("Logistic_Predictions") == 0, 0.0)
                               .otherwise(col("prediction")))

    predDF.select(outputCol, "FinalPredictions").show()
    
    regressionEvaluator = RegressionEvaluator(
    predictionCol="FinalPredictions",
    labelCol=outputCol,
    metricName="rmse")
    rmse = regressionEvaluator.evaluate(predDF)

    regressionEvaluator = RegressionEvaluator(
    predictionCol="FinalPredictions",
    labelCol=outputCol,
    metricName="r2")
    r2 = regressionEvaluator.evaluate(predDF)
      
    # Manually calculate Adjusted r2
    adjusted_r2 = adj_r2(r2, inputCols, predDF)
    
    return rmse, r2, adjusted_r2, predDF

print("** All original inputs, original output **")
inputCols = ["total_spend","total_events","purchase_events", "total_sessions", "avg_session_length", "avg_interactions_per_session", "max_interactions_per_session",
             "purchase_pct_of_total_events", "view_pct_of_total_events", "cart_pct_of_total_events","avg_purchases_per_session", "cart_events", "view_events", 
             "sessions_with_purchase", "sessions_with_cart","sessions_with_view", "pct_sessions_end_purchase", "pct_sessions_end_cart", 'sd_session_length', 
             'sd_interactions_per_session', 'sd_purchases_per_session']


outputCol = "T_total_spend"

pipeline = generateLinearPipeline(inputCols, outputCol)
pipelineModel = pipeline.fit(part_2_train)

print("Model coefficients")
print(modelInfo(inputCols, pipelineModel))

evaluationMetrics = getLinearEvaluationMetrics(pipelineModel,outputCol,part_2_test, inputCols)
print(f"RMSE is {evaluationMetrics[0]:.1f}")
print(f"R^2 is {evaluationMetrics[1]:.5f}")
print(f"Adjusted R^2 is {evaluationMetrics[2]:.5f}")

%%time
print("** Champion Model **")
inputCols = ["total_spend","purchase_events","avg_interactions_per_session", 
             "purchase_pct_of_total_events","cart_events", "view_events", 
             "sessions_with_purchase", "sessions_with_cart","pct_sessions_end_purchase", "pct_sessions_end_cart",  
             'sd_purchases_per_session']

outputCol = "T_total_spend"
# Below creation of pipeline is necessary for crossval to run.
vecAssembler = VectorAssembler(inputCols=inputCols, outputCol="unscaled_features")
# Standard Scale
ss = StandardScaler(inputCol="unscaled_features", outputCol="features")
# Select output column for linear regression
lr = LinearRegression(featuresCol="features", labelCol="T_total_spend")

pipeline = Pipeline(stages=[vecAssembler,ss, lr])
    
paramGrid = ParamGridBuilder() \
    .addGrid(lr.regParam, [0, 0.01, 0.2, 1, 10]) \
    .addGrid(lr.elasticNetParam, [0, 0.25, 0.5, 0.75, 1]) \
    .build()

crossval = CrossValidator(estimator=pipeline,
                          estimatorParamMaps=paramGrid,
                          evaluator=RegressionEvaluator().setLabelCol("T_total_spend"),
                          numFolds=4)

# Run cross-validation, and choose the best set of parameters.
cvModel = crossval.setParallelism(8).fit(part_2_train)

print("Best model coefficients")
print(modelInfo(inputCols, cvModel.bestModel))


evaluationMetrics = getLinearEvaluationMetrics(cvModel.bestModel,"T_total_spend",part_2_test,inputCols)
print(f"RMSE is {evaluationMetrics[0]:.1f}")
print(f"R^2 is {evaluationMetrics[1]:.5f}")
print(f"Adjusted R^2 is {evaluationMetrics[2]:.5f}")

print()
print(cvModel.getEstimatorParamMaps()[np.argmax(cvModel.avgMetrics)])
print()
print(cvModel.avgMetrics)
